refactor(statements): move the block pipeline out of the polygon package
Macro expansion and block reading were never Polygon's; only the final
TeX conversion is, and it becomes a parameter that stays on by default.
