+++ rbx/testdata/contests/statements_v2_polygon/A/statement/statement.rbx.tex
@@ -14,5 +14,5 @@ A single integer $n$.
 Print $n$.
 %- endblock
 %- block notes
-Be careful with $\NN$.
+Be careful with $\NN$, and with \(n \le 10\).
 %- endblock
